restructure project for better encapsulation
