+++ src/core/ray.rs
@@ -1,4 +1,4 @@
-use crate::vec3::{Point3, Vec3};
+use crate::core::vec3::{Point3, Vec3};
 
 #[derive(Default)]
 pub struct Ray {
