+++ src/io/asset_loader.rs
